Persist verified filter and general cleanup
