+++ packages/web/src/pages/search-page-v2/SearchResults.tsx
@@ -3,10 +3,10 @@ import { AllResults } from './search-results/AllResults'
 import { PlaylistResultsPage } from './search-results/PlaylistResults'
 import { ProfileResultsPage } from './search-results/ProfileResults'
 import { TrackResultsPage } from './search-results/TrackResults'
-import { useSearchCategory } from './utils'
+import { useSearchCategory } from './hooks'
 
 export const SearchResults = () => {
-  const category = useSearchCategory()
+  const [category] = useSearchCategory()
 
   switch (category) {
     case 'profiles':
