+++ packages/web/src/pages/search-page-v2/types.ts
@@ -1,4 +1,12 @@
+import {
+  SearchCategory,
+  useGetSearchResults as useGetSearchResultsApi
+} from '@audius/common/api'
+import { ID } from '@audius/common/models'
 import { IconComponent } from '@audius/harmony'
+import { Mood } from '@audius/sdk'
+
+import { categories } from './categories'
 
 export type ViewLayout = 'grid' | 'list'
 export const viewLayoutOptions: { label: string; value: ViewLayout }[] = [
@@ -29,3 +37,25 @@ export type Category = {
   filters: Filter[]
   icon?: IconComponent
 }
+
+export type MoodInfo = {
+  label: Mood
+  value: Mood
+  icon: JSX.Element
+}
+
+export type SearchResultsApiType = ReturnType<typeof useGetSearchResultsApi>
+
+export type SearchResultsType<C extends SearchCategory> = {
+  status: SearchResultsApiType['status']
+  data: C extends 'all'
+    ? {
+        users: ID[]
+        tracks: ID[]
+        playlists: ID[]
+        albums: ID[]
+      }
+    : ID[]
+}
+
+export type CategoryKey = keyof typeof categories
